+++ TAO/TAO_IDL/be/be_visitor_sequence/buffer_type.cpp
@@ -104,7 +104,7 @@ be_visitor_sequence_buffer_type::visit_string (be_string *node)
 {
   TAO_OutStream *os = this->ctx_->stream ();
 
-  if (node->width () == sizeof (char))
+  if (node->width () == (long) sizeof (char))
     {
       *os << "char *";
     }
